Implement core functionality of the module
The `Dev::ContainerizedService` module exports a (very tiny) DSL for
specifying services and setting up any environment variables. It also
exports a MAIN that starts containerized services as requseted, runs
the setup blocks, and then runs another process (which could be, for
exmaple, the service script or the cro runner) along with setting any
environment variables.